Move ExtendedArgument into its own module
