+++ src/index.ts
@@ -16,6 +16,7 @@ export * from './lib/structures/Command';
 export * from './lib/structures/CommandStore';
 export * from './lib/structures/Event';
 export * from './lib/structures/EventStore';
+export * from './lib/structures/ExtendedArgument';
 export * from './lib/structures/Precondition';
 export * from './lib/structures/PreconditionStore';
 export * from './lib/types/Enums';
